feat-#57: migrated to tailwind styles and removed unnecessary options
