+++ frontend/src/index.css
@@ -14,74 +14,29 @@
 /* React Quill Text Editor Styles */
 
 .ql-toolbar.ql-snow {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: rgb(226, 232, 240);
-  padding: 14px !important;
-  border: none !important;
-  border-radius: 0.5rem;
-  flex-wrap: wrap;
-  color: white !important;
-}
-
-.dark .ql-toolbar.ql-snow {
-  background-color: rgba(30, 41, 59, 1);
+  @apply flex flex-wrap items-center justify-center rounded-lg !border-none bg-slate-200 p-3 text-slate-50 dark:bg-dark-card;
 }
 
 .ql-editor {
-  min-height: 150px !important;
-  background-color: rgb(226, 232, 240);
-  cursor: text;
-  color: black;
-  font-size: 1rem !important;
-  border: none !important;
-  border-radius: 0.5rem;
-  margin-top: 0.5rem;
+  @apply mt-2 min-h-[150px] cursor-text rounded-lg border-none bg-slate-200 text-base text-black focus:ring-1 focus:ring-light-secondary dark:bg-dark-card dark:text-slate-50 focus:dark:ring-dark-secondary;
 }
 
-.ql-editor:focus {
-  --tw-ring-offset-shadow: var(--tw-ring-inset) 0 0 0 var(--tw-ring-offset-width)
-    var(--tw-ring-offset-color);
-  --tw-ring-shadow: var(--tw-ring-inset) 0 0 0 calc(1px + var(--tw-ring-offset-width))
-    var(--tw-ring-color);
-  box-shadow: var(--tw-ring-offset-shadow), var(--tw-ring-shadow), var(--tw-shadow, 0 0 #0000);
-  --tw-ring-opacity: 1;
-  --tw-ring-color: rgb(0 0 0 / var(--tw-ring-opacity));
-}
-
-.dark .ql-editor {
-  background-color: rgba(30, 41, 59, 1);
-  color: white !important;
-  --tw-ring-opacity: 1;
-  --tw-ring-color: rgb(255 255 255 / var(--tw-ring-opacity));
-}
-
-.ql-editor.ql-blank::before {
-  color: rgb(100, 116, 139);
-  font-style: normal;
-  font-weight: 500;
-  font-size: 0.875rem;
-  line-height: 1.25rem;
-  letter-spacing: 0.05rem;
-}
-
-.dark .ql-editor.ql-blank::before {
-  color: rgb(148, 163, 184);
+.ql-blank::before {
+  @apply text-sm !not-italic text-light-tertiary dark:text-dark-tertiary;
 }
 
 .ql-container.ql-snow {
-  border: none !important;
+  @apply border-none;
 }
 
-.dark .ql-snow * {
-  stroke: white !important;
-}
-
-.dark .ql-picker-label {
-  color: white !important;
+.ql-snow * {
+  @apply dark:stroke-slate-50;
 }
 
 .ql-picker-label {
-  color: black !important;
+  @apply !text-black dark:!text-slate-50;
+}
+
+.ql-picker-options {
+  @apply !bg-slate-200 text-light-secondary dark:!bg-dark-card dark:text-slate-50;
 }
